added update function for updating the cabin details

+++ src/services/apiCabins.js
@@ -11,34 +11,47 @@ export async function getCabins() {
   return data;
 }
 
-export async function createCabin(newCabin) {
-  //Create the file path
+export async function createEditCabin(newCabin, id) {
+  console.log(newCabin);
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
 
-  //Upload the image itself
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
+
   if (error) {
-    console.log(error);
-    throw new Error("Cabin cannot be created.");
+    console.error(error);
+    throw new Error("Cabin could not be created");
   }
+
+  // 2. Upload image
+  if (hasImagePath) return data;
 
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
-    console.log(storageError);
+    console.error(storageError);
     throw new Error(
-      "Cabin image could not be uploaded and the cabin was not created."
+      "Cabin image could not be uploaded and the cabin was not created"
     );
   }
